Rename fields to be consistent with rfc3501
The fields have slightly different names in rfc3501. For example, the
"\" followed by an atom are called name attributes and not flags when
they concern a mailbox. The flags name is used for messages. This
made it easier for me to see what in the standard the code corresponds
to.

From [rfc3501 section 6.3.8](https://datatracker.ietf.org/doc/html/rfc3501#section-6.3.8):

> The LIST command returns a subset of names from the complete set
> of all names available to the client.  Zero or more untagged LIST
> replies are returned, containing the name attributes, hierarchy
> delimiter, and name; see the description of the LIST reply for
> more detail.